Add mutex around Shell
Prior to this change, shell commands could run in parallel, which was particularly messy for Sudo since we would end up with multiple sudo password prompts.

This change adds a mutex in Shell so only one shell command can execute at a time. This tidies up the terminal output a bit. We still have a bit of a synchronisation problem - sudo prompts can get written in the middle of output. This is probably going to be solved by just not printing shell output when we don't need to!